moving `XYZ::dominant_wavelength` to XYZ::dominant.rs submodule
